requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2021.10.8
cffi==1.15.0
Django==4.2.4
django-bootstrap3==23.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-taggit==4.0.0
django-tinymce==3.6.1
djangorestframework==3.14.0
greenlet==2.0.2
misaka==2.1.1
Pillow==10.0.0
pycparser==2.21
pytz==2023.3
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
wincertstore==0.2
